# backend/torneosService/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Torneos Service - Gestion Deportiva UCB
  description: API documentada solo con los endpoints usados por el frontend de gestion deportiva.
  version: 1.0.0
servers:
  - url: /
    description: Servidor local del microservicio de torneos
...
paths:
...
  /reserva:
    get:
      tags: [Reservas]
      summary: Listar reservas
      parameters:
        - name: fecha
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: equipoId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Reservas registradas
    post:
      tags: [Reservas]
      summary: Crear reserva
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservaInput"
      responses:
        "201":
          description: Reserva creada
  /reserva/disponibilidad/{canchaId}/{fecha}:
    get:
      tags: [Reservas]
      summary: Consultar disponibilidad de cancha
      parameters:
        - name: canchaId
          in: path
          required: true
          schema:
            type: integer
        - name: fecha
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Horarios disponibles
  /reserva/{id}:
    get:
      tags: [Reservas]
      summary: Obtener reserva por ID
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Reserva encontrada
    patch:
      tags: [Reservas]
      summary: Actualizar o cancelar reserva
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservaInput"
      responses:
        "200":
          description: Reserva actualizada
    delete:
      tags: [Reservas]
      summary: Eliminar reserva
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Reserva eliminada
  /torneo:
    get:
      tags: [Torneos]
      summary: Listar torneos
      responses:
        "200":
          description: Torneos registrados
    post:
      tags: [Torneos]
      summary: Crear torneo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TorneoInput"
      responses:
        "201":
          description: Torneo creado
  /torneo/{id}:
    get:
      tags: [Torneos]
      summary: Obtener torneo por ID
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Torneo encontrado
    patch:
      tags: [Torneos]
      summary: Actualizar torneo
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TorneoInput"
      responses:
        "200":
          description: Torneo actualizado
    delete:
      tags: [Torneos]
      summary: Eliminar torneo
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Torneo eliminado
  /torneo-equipo:
    post:
      tags: [Torneos]
      summary: Agregar equipo a torneo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TorneoEquipoInput"
      responses:
        "201":
          description: Equipo agregado al torneo
  /torneo-equipo/{torneoId}/{equipoId}:
    delete:
      tags: [Torneos]
      summary: Quitar equipo de torneo
      parameters:
        - name: torneoId
          in: path
          required: true
          schema:
            type: integer
        - name: equipoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipo removido del torneo
...
  /fixture/torneo/{torneoId}:
    get:
      tags: [Fixture]
      summary: Obtener partidos por torneo
      parameters:
        - name: torneoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Partidos del torneo
...
  /partidos/torneo/{idTorneo}:
    get:
      tags: [Partidos]
      summary: Obtener todos los partidos de un torneo
      parameters:
        - name: idTorneo
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Partidos del torneo
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ReservaInput:
      type: object
      properties:
        cancha_id:
          type: integer
          example: 1
        equipo_id:
          type: integer
          example: 2
        fecha:
          type: string
          format: date
          example: "2026-07-05"
        hora_inicio:
          type: string
          example: "14:00"
        hora_fin:
          type: string
          example: "16:00"
        estado:
          type: string
          example: confirmada
        observaciones:
          type: string
          example: Entrenamiento
    TorneoInput:
      type: object
      properties:
        nombre:
          type: string
          example: Torneo Intercarreras 2026
        tipo:
          type: string
          example: Interno
        estado:
          type: string
          example: planeado
        disciplina_id:
          type: integer
          example: 1
        fecha_inicio:
          type: string
          format: date
          example: "2026-07-01"
        fecha_fin:
          type: string
          format: date
          example: "2026-07-30"
        imagen_url:
          type: string
          example: https://example.com/torneo.jpg
    TorneoEquipoInput:
      type: object
      required: [torneo_id, equipo_id]
      properties:
        torneo_id:
          type: integer
          example: 1
        equipo_id:
          type: integer
          example: 2
...